collectors, targets and parsers
To simplify targeted refresh, we don't differentiate between targeted refresh
and full refresh anymore, but a full refresh is simply the manager as a
target. So in this ansible example, it's automation_manager.

Collaborators

Inventory/Collector

A collector acts as a proxy to the providers inventory. It's task is to
provide methods that collect this data. It's base class is initialized
with manager and target, so it can make use of with_provider_connection
or restrict collected data to the target scope. The target in this case
is the ActiveRecord instance of the refresh we are targeting. So in a
full refresh it's the same as the manager. If we target a Host, the
target is the Host instance under the manager

Inventory/Target

This holds all InventoryCollections and makes them accessible via simple
methods. Thats needed for cross referencing InventoryObjects via
lazy_find

Inventory/Parser

The base class holds the above collector and inventory/target. The entry
method is parse which fills the inventory collections by making use of
the collector

Inventory

ManagerRefresh::Inventory takes care of wiring up the above 3 objects:
Collector, Target and Parser. It's constructed in the platform and the
instances of the 3 objects are constructed in the platform as well. It's
all happening in ems_refresher_mixin mainly in parse_targeted_inventory.
So ideally this is hidden from the provider author.

What's required from a provider author?

Basically the author provides these 3 classes, which inherit from their
base classes in the platform:

<provider>/inventory/collectors/<target_name.rb>
<provider>/inventory/parsers/<target_name.rb>
<provider>/inventory/target/<target_name.rb>

In case the classes are present, targeted refresh JustWorks ™️


(transferred from ManageIQ/manageiq@ec59992ccf247e328098dfd91240a3dbdde33834)