Adding many models for EmbeddedAnsible provider
WOW!  So many changes here.

This is the first attempt to create a way to differentiate the
AnsibleTower models from the EmbeddedAnsible models.

The point behind this differentiation is to allow querying done via the
MIQ API and UI to filter based on the Embedded vs. External nature of
each.

Specifically, one should be able to ask for
`ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScript.all`
and not get any configuration script from any `AnsibleTower` providers,
because `AnsibleTower` providers should be strictly
`ExternalAutomationManagers`.

This definitely feels like a lot of convoluted hierarchical organization
that's going to hurt in the future.

TODO:  OMG, draw a picture of this!


(transferred from ManageIQ/manageiq@be9b4d0fc701240ce1387b57d2fb68c2c96955ec)