Remove dependency on custom have_attributes
In core we have a custom `have_attributes` matcher that overrides the
normal RSpec one. Among other silly things, it allows you to pass a time
object and assume that you want to approximate it (which was yet another
custom matcher, too...).

We are removing these matchers to reduce confusion and provide the
actual RSpec ones which are more feature rich anyway.

Things like time approximation can and should be done in the caller
doing something like this (see code).