Wait for project update to finish
It may not be possible to create a job template right after creating a
project that the template uses. Add a try-wait loop that ensures that
the project has been created in Tower and it successfully finished
cloning.

In case of clone failure, raise an error. If the project is not ready
yet, wait a bit and try again. The number of tries is limited, raising
an error on depletion.

Extract the del_object sleep as a class constant for symmetry.