Collect and use variable Tower data
Each time objects are populated in a Tower instance, some of their
properties (ids, timestamps…) change. The same applies to objects counts
that can differ from the previous state. This required manual update of
expectation in specs when re-recording VCR cassettes.

Collect these changing data during their population in Tower instance to
a YAML file. Use this information in specs instead of magic literals.
This allows to re-record the cassettes without breaking the tests and
removes the need to manually edit the expectations in the specs.