Collect variable Tower data upon population
Each time objects are populated in a Tower instance, some of their
properties (ids, timestamps…) change. The same applies to objects counts
that can differ from the previous state. This information is used then
in the specs.

Collect these changing data during their population in Tower instance to
a YAML file. This allows to fill the necessary information in the
specs after re-recording the cassettes more easily. Eventually it
could be used directly from the tests too.